+++ src/test/java/com.xebialabs.xlr.xlrelease.test/LoginTest.java
@@ -29,6 +29,48 @@ public class LoginTest extends BaseTest {
         SharedConfigurationPropertiesPage.checkSharedConfigurationHeader("ServiceNow");
     }
 
+    @Test
+    public void SaveConfiguration(){
+        MainMenu.clickMenu("Settings");
+        SubMenu.clickSubMenu("Shared configuration");
+        SharedConfigurationPage.openSharedConfiguration("ServiceNow: Server");
+        SharedConfigurationPropertiesPage.checkSharedConfigurationHeader("ServiceNow");
+        SharedConfigurationPropertiesPage.setEditFieldBySequence(1,"ServiceNow Config 1");
+        SharedConfigurationPropertiesPage.setEditFieldBySequence(2,"ServiceNow url");
+        SharedConfigurationPropertiesPage.clickButtonByText("Save");
+        SharedConfigurationPropertiesPage.isNewConfigurationSaved().isSharedConfigurationPageVisible("ServiceNow: Server");
+    }
+
+    @Test
+    public void TestInvalidConfiguration(){
+        MainMenu.clickMenu("Settings");
+        SubMenu.clickSubMenu("Shared configuration");
+        SharedConfigurationPage.openSharedConfiguration("ServiceNow: Server");
+        SharedConfigurationPropertiesPage.checkSharedConfigurationHeader("ServiceNow");
+        SharedConfigurationPropertiesPage.setEditFieldBySequence(1,"ServiceNow Config 2");
+        SharedConfigurationPropertiesPage.setEditFieldBySequence(2,"ServiceNow Url");
+        SharedConfigurationPropertiesPage.clickElementById("authenticationMethod"); // clicking the element so that select field will be visible on next step
+        SharedConfigurationPropertiesPage.setOptionFromSelectFieldBySequence(1,"None");
+        SharedConfigurationPropertiesPage.clickButtonByText("Test");
+        SharedConfigurationPropertiesPage.checkConnectionStatusShouldContain("Can't connect");
+    }
+
+    @Test
+    public void TestValidConfiguration(){
+        MainMenu.clickMenu("Settings");
+        SubMenu.clickSubMenu("Shared configuration");
+        SharedConfigurationPage.openSharedConfiguration("ServiceNow: Server");
+        SharedConfigurationPropertiesPage.checkSharedConfigurationHeader("ServiceNow");
+        SharedConfigurationPropertiesPage.setEditFieldBySequence(1,"ServiceNow Config 3");
+        SharedConfigurationPropertiesPage.setEditFieldBySequence(2, System.getenv("SERVICENOW_URL"));
+        SharedConfigurationPropertiesPage.clickElementById("authenticationMethod"); // clicking the element so that select field will be visible on next step
+        SharedConfigurationPropertiesPage.setOptionFromSelectFieldBySequence(1,"Basic");
+        SharedConfigurationPropertiesPage.setEditFieldBySequence(7, System.getenv("SERVICENOW_USERNAME"));
+        SharedConfigurationPropertiesPage.setEditFieldBySequence(8, System.getenv("SERVICENOW_PASSWORD"));
+        SharedConfigurationPropertiesPage.clickButtonByText("Test");
+        SharedConfigurationPropertiesPage.checkConnectionStatusShouldContain("ServiceNow: Server is available");
+    }
+
     @AfterMethod
     public void logout(){
         System.out.println("called after method");