+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Improvement
+- Updated nebula release plugin to v6.3.5
+- Updated xl-test-api to v0.0.4
+- Added test connection tests for shared configuration
+
 ## v4.1.14
 
 ### Fixed